refactorising into a dedicated module next to the module manager
as its kind of managing modules!
